[fromlist] wpa_supplicant: Add an option to disable SNR capping
When selecting a network the default behaviour of WPA supplicant is to
prefer higher throughput, it does this by capping the SNR.

But in certain environments, reliability is important over throughput
and choosing a lower SNR (thought it is greater than "Great SNR") might
be sub-optimal.

Introduce a configuration option to choose the two options (throughput
or reliability).

Upstream PR: http://lists.infradead.org/pipermail/hostap/2024-March/042483.html

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>