STAUT fails to connect to Ex-AP with enterprise security.
[Description]
DUT as STA fails to connect to Ex-AP with enterprise security after uAP is started and stopped in enterprise security.

[Root Cause]
After uAP is started stopped, tls_ctx_global.tls_conf is not freed and it is still not NULL.
If tls_ctx_global.tls_conf is not NULL, it will choose SERVER ROLE when wlan connect is provided again.
STA connection fail for endpoint as it is set to SERVER ROLE (should be CLIENT ROLE) due to which it respond with wrong EAP packet to peer.
The hello client packet is without client content due to which DUT receives deauth by EX-AP.
The reason that tls_ctx_global.tls_conf is not freed and is not set to NULL is because tls_deinit() is returned early when tls_ctx_global.refcnt is not 0.
Due to which it will not call tls_conf_deinit() and will not set tls_ctx_global.tls_conf NULL.

[Fix]
tls_conf_deinit() is now called before tls_ctx_global.refcnt check and tls_ctx_global.tls_conf has it's own refcnt.

[Testing]
DUT as STA is now able to connect to Ex-AP with enterprise security after uAP is started and stopped in enterprise security.

[Open Issues/Todo]
None.

[Reviewers]
Gannicus Guo, Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Keval Thakker, Sanjay Yadav, Pankaj Razdan.

Change-Id: a9f1cd08-c9d3-11ef-af8c-4bcc7d01a01d

Signed-off-by: Mayur Mansharamani <mayur.mansharamani@nxp.com>