Fix hang issue found in Enterprise stress tests
[Description]
Hang issue was found when doing Enterprise stress tests of uAP.

[Root Cause]
For Enterprise mode, supplicant will register port timers tick to
eloop and it will timeout every 1 second. Right before hang issue
happened, the cancelation of this timer from eloop was failed and the
timer was added back to eloop after uap was disabled. When this timer
was timeout again, the input paramter timeout_ctx was invalid and caused
hardfault when supplicant task is trying to use this invalid parameter.
The cancelation of this timer is called in wlcmgr task while the timer
handler is called in supplicant task. This is a corner case that the
timer handler was called first. Before the execution of handler is done,
cancelation is triggered. The cancelation was executed right away
because wlcmgr task has higher priority of supplicant task.Since the
timer handler is already running, the timer node was dequeued from eloop.timeout
and couldn't be found when doing cancelation. After cancelation is done
and supplicant task has chance to run, the timer will be added back to
eloop.timeout before exit. The timer kept timeout after uAP is stopped
and called invalid parameter and finally caused hang issue.

[Fix]
Fetch wpa_supplicant_mutex in eloop before calling timeout handler. In
this case, wlcmgr task can't get the mutex and have to wait for eloop to
finish. Release the mutex after eloop is done.

[Testing]
Verification is ok. No hang was found in stress tests.

[Reviewers]
Gannicus Guo

[Port proposal]
non-general

Change-Id: If407f0af36fcc1379c8b361a4b4ca00928c90f6f
Signed-off-by: Hui Bai <hui.bai@nxp.com>