Add support for station network priority groups
[Description]
[Supplicant] STA should connect another available AP in networklist automatically when current AP is unavailable.

[Root Cause]
When connecting to the specified AP, STA marks all other networks as disabled
and does not use the priority groups mechanism.

[Fix]
a. Add the priority parameter in the network profile to set the priority of each network.
b. When a priority group is set, networks will be selected based on priority.
   By default, all networks will get same priority group (0).
   The priority groups will be iterated in decreasing priority (i.e., the larger the
   priority value, the sooner the network is matched against the scan results).
   Within each priority group, networks will be selected based on security policy, signal
   strength, etc.

[Testing]
Test pass.

[Reviewers]
Gannicus Guo, Maochen Wang, Hui Bai, Supriya Sumeet Inamdar

Change-Id: I9d3d341d38e941dfd62a6ec2adbd776515edff6f
Signed-off-by: Gang Li <gang.li_1@nxp.com>